+++ Python/server/__init__.py
@@ -67,6 +67,7 @@ def bootstrap():
     from server import testing_validation_tools     # noqa: F401  Sub-batch W
     from server import data_table_extension_tools   # noqa: F401  Sub-batch X
     from server import metasound_tools              # noqa: F401  Sub-batch Y
+    from server import sequencer_extension_tools    # noqa: F401  Sub-batch Z
     from server import packaging_tools             # noqa: F401
     from server import vertical_test_tools        # noqa: F401